code/ticket-presidio/requirements.txt
mysql-connector-python
pandas
spacy